+++ packages/web/src/react-query/router/index.ts
@@ -1,2 +1,2 @@
 export * from "./use-get-swap-fee";
-export * from "./use-estimate-swap";
+export * from "./use-get-routes";
